Refactor background script into smaller files
