+++ src/background/actions/getHighlights.js
@@ -1,0 +1,7 @@
+import { executeInCurrentTab } from '../utils.js';
+
+function getHighlights() {
+    return executeInCurrentTab({ file: 'src/contentScripts/getHighlights.js' });
+}
+
+export default getHighlights;
