+++ src/background/actions/index.js
@@ -1,0 +1,11 @@
+export { default as highlightText } from './highlightText.js';
+export { default as toggleHighlighterCursor } from './toggleHighlighterCursor.js';
+export { default as removeHighlights } from './removeHighlights.js';
+export { default as removeHighlight } from './removeHighlight.js';
+export { default as showHighlight } from './showHighlight.js';
+export { default as getHighlights } from './getHighlights.js';
+export { default as getLostHighlights } from './getLostHighlights.js';
+export { default as changeColor } from './changeColor.js';
+export { default as editColor } from './editColor.js';
+export { default as getCurrentColor } from './getCurrentColor.js';
+export { default as getColorOptions } from './getColorOptions.js';
